nprs_pks: Move signature handling logic out of the nrpspredictor module
