+++ antismash/modules/nrps_pks/nrps_predictor.py
@@ -18,15 +18,7 @@ from antismash.config import ConfigType
 from antismash.detection.nrps_pks_domains import ModularDomain
 
 from .data_structures import Prediction
-
-REF_SEQUENCE = "P0C062_A1"
-A34_POSITIONS_FILENAME = path.get_full_path(__file__, "external", "NRPSPredictor2", "A34positions.txt")
-APOSITION_FILENAME = path.get_full_path(__file__, "external", "NRPSPredictor2", "Apositions.txt")
-KNOWN_CODES = path.get_full_path(__file__, "knowncodes.fasta")
-ILLEGAL_CHARS = "!@#$%^&*(){}:\"<>?/.,';][`~1234567890*-+-=_\\|"
-ADOMAINS_FILENAME = path.get_full_path(__file__, "external", "NRPSPredictor2", "A_domains_muscle.fasta")
-START_POSITION = 66
-
+from .signatures import get_a_dom_signatures
 
 def _build_stach_codes() -> Dict[str, Set[str]]:
     """ Builds a mapping of Stachelhaus prediction to code from NRPSPredictor2's
@@ -207,102 +199,6 @@ class PredictorSVMResult(Prediction):
                                   json["stachelhaus_match_count"])
 
 
-def read_positions(filename: str, start_position: int) -> List[int]:
-    """ Loads positions from a tab-separated file. Positions are relative to the start_position.
-
-        Arguments:
-            filename: the path to the file containing the positions
-            start_position: a relative start position to adjust all positions by
-
-        Returns:
-            a list of ints, one for each position found in the file
-    """
-    with open(filename, "r") as data:
-        text = data.read().strip()
-    results = []
-    for i in text.split("\t"):
-        results.append(int(i) - start_position)
-    return results
-
-
-def build_position_list(positions: List[int], reference_seq: str) -> List[int]:
-    """ Adjusts a list of positions to account for gaps in the reference sequence
-
-        Arguments:
-            positions: a list of ints that represent positions of interest in
-                       the reference sequence
-            reference_seq: the (aligned) reference sequence
-
-        Returns:
-            a new list of positions, each >= the original position
-    """
-    poslist = []
-    position = 0
-    for i, ref in enumerate(reference_seq):
-        if ref != "-":
-            if position in positions:
-                poslist.append(i)
-            position += 1
-    return poslist
-
-
-def verify_good_sequence(sequence: str) -> bool:
-    """ Ensures a sequence is valid """
-    for char in ILLEGAL_CHARS:
-        if char in sequence:
-            return False
-    return True
-
-
-def extract(sequence: str, positions: List[int]) -> str:
-    """ Extracts a signature from an aligned sequence based on the provided
-        positions. Accounts for gaps by looking behind or, if behind is already
-        in the position list, ahead.
-
-        Arguments:
-            sequence: the aligned sequence to extract a signature from
-            positions: the list of positions within the sequence to use
-
-        Returns:
-            the extracted signature as a string
-    """
-    seq = []
-    for position in positions:
-        aa = sequence[position]
-        if aa == "-":
-            if position - 1 not in positions:
-                aa = sequence[position - 1]
-            elif position + 1 not in positions:
-                aa = sequence[position + 1]
-        seq.append(aa)
-    return "".join(seq)
-
-
-def get_34_aa_signature(domain: ModularDomain) -> str:
-    """ Extract 10 / 34 AA NRPS signatures from A domains """
-    assert " " not in domain.get_name()
-    assert verify_good_sequence(domain.translation)
-    # Run muscle and collect sequence positions from file
-    alignments = subprocessing.run_muscle_single(domain.get_name(), domain.translation, ADOMAINS_FILENAME)
-
-    domain_alignment = alignments[domain.get_name()]
-    reference_alignment = alignments[REF_SEQUENCE]
-
-    positions = read_positions(APOSITION_FILENAME, START_POSITION)
-    # Count residues in ref sequence and put positions in list
-    poslist = build_position_list(positions, reference_alignment)
-
-    # Extract positions from query sequence
-    query_sig_seq = extract(domain_alignment, poslist)
-    # Add fixed lysine 517
-    query_sig_seq += "K"
-
-    # repeat with 34 AA codes
-    angpositions = read_positions(A34_POSITIONS_FILENAME, START_POSITION)
-    poslist = build_position_list(angpositions, reference_alignment)
-
-    return extract(domain_alignment, poslist)
-
 
 def read_output(lines: List[str]) -> Dict[str, Prediction]:
     """ Converts NRPSPredictor2 output lines to Predictions
@@ -346,7 +242,7 @@ def run_nrpspredictor(a_domains: List[ModularDomain], options: ConfigType) -> Di
     output_filename = "svm_output.txt"
     bacterial = "1" if options.taxon == "bacteria" else '0'
 
-    signatures = [get_34_aa_signature(a_domain) for a_domain in a_domains]
+    signatures = [get_a_dom_signatures(a_domain)[1] for a_domain in a_domains]
 
     with TemporaryDirectory(change=True):
         # Get NRPSPredictor2 code predictions, output sig file for input for NRPSPredictor2 SVMs
